remove logs, create utils file
