+++ apps/consent-ui/src/components/Header/LanguageToggle.tsx
@@ -24,33 +24,9 @@ import clsx from 'clsx';
 
 import styles from '@/components/Button/Button.module.scss';
 import { ValidLanguage } from '@/i18n';
-import routesByLocale from '@/i18n/routes/routesByLocale.json';
-import { RouteName, RouteNameEnum } from '@/components/Link/types';
 
 import LocalizedLink from '../Link/LocalizedLink';
-
-import { getUnselectedLang } from '.';
-
-function getRouteNameByPath(routeObj: { [k: string]: string }, value: string): string | undefined {
-	const keys = Object.keys(routeObj);
-	return keys.find((key: string) => routeObj[key] === value);
-}
-
-const findLinkNameByPath = (path: string, lang: ValidLanguage): RouteName => {
-	if (!path) {
-		return 'home';
-	}
-	const pathSegments = path.split('/');
-	const newPath = pathSegments.slice(2).join('/');
-	const result = getRouteNameByPath(routesByLocale[lang], `/${newPath}`);
-	try {
-		const validRoute = RouteNameEnum.parse(result);
-		return validRoute;
-	} catch (e) {
-		console.error(`Invalid route name: ${result}`);
-		return 'home';
-	}
-};
+import { findLinkNameByPath, getUnselectedLang } from '../Link/utils';
 
 function LanguageToggle({ lang, children }: { lang: ValidLanguage; children: ReactNode }) {
 	const langToSelect = getUnselectedLang(lang);
